requirements.txt
pyaml
matplotlib
pymongo
numpy
